New features. Removed a function
All functions to compute the number of crossings now have a '_list' version.
Added enumeration for the different algorithms for the computation of C.
Used the enumeration to make a function 'n_crossings' and 'n_crossings_list'.
All plain 'enum' are now 'enum class'.

Removed function to compute C using the set Q.