Slight modification of Wright's algorithm.
Wright's algorithm: is the algorithm used for generating free unlabelled trees
in amortized constant time.

The modification simply shifts the contents of the vectors m_L and m_W of the
class 'all_ulab_free_trees' one position to the right. Recall that m_L is the
level sequence of the tree that has to be generated. This modification affects
the behaviour of the conversion of a level sequence of a tree into an actual
tree data structure.