Fixed bugs
The enumeration of projective arrangements was flawed. The intervals of
vertices may not have been initialised properly: they need to be
initialised with a list of indices sorted increasingly so as to be able
to enumerate all the permutations of each interval.

However, so far it worked correctly since the trees that I used
always had vertex '0' as their root, were normalised. For this reason,
the intervals were initialised with vertices sorted increasingly.