Removed empty constructors. Removed init() method. Added reset() method.
The tree generators can no longer be instantiated with an empty
constructor, and thus must be initialised with some value (the number of
vertices). The method init() has also been removed from every generator
class.

Therefore, this:
	all_ulab_free_trees Gen;
	Gen.init(n);
becomes:
	all_ulab_free_trees Gen(n);

Added method reset() to the classes for exhaustive generation of trees
(not random generation).