Add documentation to the Python module files
For a better experience when coding with software like Spyder.

The files added are a script to automatize the generation of the python
module (or interface) files. Makefiles have been modified to be able to
generate only the .py files.

The .py files need postprocessing for a clean up of notation extraneous
to programmers not familiar to C++.