Improved error messages. Added detection of more errors
Error messages contain a prefix indicating whether the "error" is an actual Error or a Warning.

Return a warning if empty lines were found.
Return an error if isolated vertices were found in the graph.
Return an error if the number of edges does not match the number of vertices minus 1.