Minor changes after improvement of YS's algorithm
JL improved the algorithm by removing the need to flip the arrangement
in the commit previous to this one.

In this commit, some comments have been removed, some commented code has
also been removed, and lastly, removed some 'if' statements and use the
operator '?' to "replace them". Efficiency has not changed with respect
to the previous commit.