Added HS-like algorithms for projective/planar minimum D
Coded Hochberg and Stallmann's algorithms (with the correction indicated
in our paper) to solve the minimum planar and projective arrangement
problems.

This needed some refactoring of the code. Look at lal/internal/linarr/Dmin_utils.hpp
This file now contains the functions to compute the rooted adjacency for
free trees and for rooted trees (the latter is optimised -- does less
computations than in the function for free trees). It also contains the
algorithms that arrange the trees using intervals and using
displacements with respect to the root. It also contains some
definitions common to all these algorithms.