Refactored code
Moved much code into header files into the lal::detail namespace.

Added a template parameter for the linear arrangement type in all the
functions that calculate the number of crossings.

In treebank processor, call the functions that calculate values
directly with the identity arrangement to save as much time as possible.