Fixed minor issues in CMake
The function 'check_cxx_compiler_flag(<flag> <var>)' caches the result
into variable '<var>': if <var> already contains a (true) result, the
function never checks again, even if the flag is different. So far the
function was called with variable 'CXX_HAS_FLAG': regardless of the
flags the function was called with were being different, the check was
performed only once, which was the first time it was called.

This has been addressed and the check is performed for every flag now.