Relaxed usage of add_edge_bulk and finish_bulk_add + new method
The method add_edge_bulk() does not impose that all edges have to be
eventually added. The method finish_bulk_add() updates the internal data
structures of the graphs but does not require the graph to be complete
in any sense.

Added a new method (finish_bulk_add_complete) only for trees that
constructs the Union-Find data structure with a simple method. This does
assume that the tree is complete (hence '_complete').